Moved SharedCombatModeSystem changes to RMCCombatModeSystem
