+++ Content.Shared/CombatMode/SharedCombatModeSystem.cs
@@ -1,5 +1,3 @@
-using Content.Shared._RMC14.Xenonids.Evolution;
-using Content.Shared._RMC14.Xenonids.Plasma;
 using Content.Shared.Actions;
 using Content.Shared.Mind;
 using Content.Shared.MouseRotator;
@@ -24,7 +22,6 @@ public abstract class SharedCombatModeSystem : EntitySystem
         SubscribeLocalEvent<CombatModeComponent, MapInitEvent>(OnMapInit);
         SubscribeLocalEvent<CombatModeComponent, ComponentShutdown>(OnShutdown);
         SubscribeLocalEvent<CombatModeComponent, ToggleCombatActionEvent>(OnActionPerform);
-        SubscribeLocalEvent<CombatModeComponent, XenoChangingPrototypeEvent>(OnXenoChangingPrototype);
     }
 
     private void OnMapInit(EntityUid uid, CombatModeComponent component, MapInitEvent args)
@@ -50,16 +47,6 @@ public abstract class SharedCombatModeSystem : EntitySystem
 
         var msg = component.IsInCombatMode ? "action-popup-combat-enabled" : "action-popup-combat-disabled";
         _popup.PopupClient(Loc.GetString(msg), args.Performer, args.Performer);
-    }
-
-    private void OnXenoChangingPrototype(Entity<CombatModeComponent> entity, ref XenoChangingPrototypeEvent args)
-    {
-        var compName = EntityManager.ComponentFactory.GetComponentName<CombatModeComponent>();
-        if (args.NewComponents.TryGetComponent(compName, out var c) && c is CombatModeComponent { } newComponent)
-        {
-            // we have combat mode and we will continue to have it. just don't do anything in this case.
-            args.AdditionalExclusions.Add(compName);
-        }
     }
 
     public void SetCanDisarm(EntityUid entity, bool canDisarm, CombatModeComponent? component = null)
